# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Marmara Üniversitesi - Açık Kaynak Kodlu Yazılımlar",
    "description": "Rest API Dokümantasyonu"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "API School",
      "description": "Node.js ile REST API implementasyonu"
    }
  ],
  "schemes": ["http", "https"],
  "paths": {
    "/students": {
      "post": {
        "tags": ["API School"],
        "summary": "Create a new student",
        "description": "CREATE - Create a new resource",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "tags": ["API School"],
        "description": "LIST - Get all resources",
        "summary": "Get all students",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/students/{id}": {
      "get": {
        "tags": ["API School"],
        "description": "READ - Get a resource",
        "summary": "Get a student",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": ["API School"],
        "description": "UPDATE - Update a resource",
        "summary": "Update a student",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": ["API School"],
        "description": "// DELETE - Delete a resource",
        "summary": "Delete a student",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}